requirements.txt
supabase==2.10.0
boto3==1.34.0
python-dotenv==1.0.1
click==8.1.7
pandas==2.2.0
rich==13.7.0
jinja2==3.1.3
python-dateutil==2.9.0
